+++ e2e-tests/dragent/crewai/register.py
@@ -59,9 +59,9 @@ class CrewaiAgentConfig(AgentBaseConfig, name="crewai_agent"):
 async def crewai_agent(config: CrewaiAgentConfig, builder: Builder) -> AsyncGenerator:
     from datarobot_genai.core.mcp import MCPConfig
     from datarobot_genai.crewai.mcp import mcp_tools_context
+    from datarobot_genai.dragent.context import extract_authorization_from_context
+    from datarobot_genai.dragent.context import extract_datarobot_headers_from_context
     from datarobot_genai.dragent.frontends.converters import aggregate_dragent_event_responses
-    from datarobot_genai.nat.helpers import extract_authorization_from_context
-    from datarobot_genai.nat.helpers import extract_datarobot_headers_from_context
     from nat.builder.function_info import FunctionInfo
     from nat.builder.function_info import Streaming
