+++ e2e-tests/dragent/llamaindex/register.py
@@ -48,10 +48,10 @@ class LlamaindexAgentConfig(AgentBaseConfig, name="llamaindex_agent"):
 )
 async def llamaindex_agent(config: LlamaindexAgentConfig, builder: Builder) -> AsyncGenerator:
     from datarobot_genai.core.mcp import MCPConfig
+    from datarobot_genai.dragent.context import extract_authorization_from_context
+    from datarobot_genai.dragent.context import extract_datarobot_headers_from_context
     from datarobot_genai.dragent.frontends.converters import aggregate_dragent_event_responses
     from datarobot_genai.llama_index.mcp import mcp_tools_context
-    from datarobot_genai.nat.helpers import extract_authorization_from_context
-    from datarobot_genai.nat.helpers import extract_datarobot_headers_from_context
     from nat.builder.function_info import FunctionInfo
 
     from dragent.llamaindex.myagent import MyAgent
