+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "datarobot-genai"
-version = "0.24.1"
+version = "0.25.0"
 description = "Generic helpers for GenAI"
 readme = "README.md"
 requires-python = ">=3.11, <3.14"
@@ -23,8 +23,8 @@ okta_cross_app_access = "datarobot_genai.dragent.plugins.okta_a2a_auth"
 per_user_tool_calling_agent = "datarobot_genai.dragent.plugins.per_user_tool_calling_agent"
 streaming_memory_agent = "datarobot_genai.dragent.plugins.streaming_memory_agent"
 datarobot_otel_conventions_middleware = "datarobot_genai.dragent.plugins.datarobot_otel_conventions_middleware"
-datarobot_mcp_client = "datarobot_genai.nat.datarobot_mcp_client"
-datarobot_auth_provider = "datarobot_genai.nat.datarobot_auth_provider"
+datarobot_mcp_client = "datarobot_genai.dragent.plugins.datarobot_mcp_client"
+datarobot_auth_provider = "datarobot_genai.dragent.plugins.datarobot_auth_provider"
 datarobot_moderation_middleware = "datarobot_genai.dragent.plugins.datarobot_moderation_middleware"
 datarobot_dragent_normalization = "datarobot_genai.dragent.plugins.datarobot_dragent_normalization"
 datarobot_otelcollector = "datarobot_genai.dragent.plugins.datarobot_otelcollector"
