+++ src/datarobot_genai/dragent/plugins/llm_clients.py
@@ -57,7 +57,6 @@ EXCLUDE_FIELDS = {
     "type",
     "thinking",
     "reasoning",
-    "headers",
     "api_type",
     "llm_deployment_id",
     "nim_deployment_id",
@@ -156,11 +155,7 @@ async def datarobot_llm_deployment_langchain(
 
     config = prepare_llm_parameters(llm_config)
 
-    context_headers = extract_headers_from_context(["X-DataRobot-Identity-Token"])
-    if llm_config.headers:
-        context_headers = {**context_headers, **llm_config.headers}
-
-    config["extra_headers"] = context_headers
+    config["extra_headers"] = extract_headers_from_context(["X-DataRobot-Identity-Token"])
 
     client = get_datarobot_deployment_llm(
         llm_config.llm_deployment_id, llm_config.model_name, parameters=config
@@ -208,8 +203,7 @@ async def datarobot_llm_component_langchain(
     if llm_type == LLMType.GATEWAY:
         client = get_datarobot_gateway_llm(llm_config.model_name, config)
     elif llm_type == LLMType.DEPLOYMENT:
-        if llm_config.headers:
-            config["extra_headers"] = llm_config.headers
+        config["extra_headers"] = extract_headers_from_context(["X-DataRobot-Identity-Token"])
         client = get_datarobot_deployment_llm(
             llm_config.llm_deployment_id,  # type: ignore[arg-type]
             llm_config.model_name,
