+++ tests/dragent/plugins/fixtures/moderation_integration/moderation_config.yaml
