+++ tests/dragent/plugins/fixtures/moderation_model_prompt_replace/moderation_config.yaml
