+++ tests/dragent/plugins/fixtures/moderation_model_response_replace/moderation_config.yaml
