+++ tests/dragent/plugins/fixtures/moderation_response_token_length_block/moderation_config.yaml
