+++ tests/dragent/plugins/fixtures/workflow_with_a2a.yaml
