+++ tests/dragent/plugins/fixtures/workflow_with_datarobot_user_mcp_xaa_client.yaml
