+++ tests/dragent/plugins/test_a2a.py
@@ -17,11 +17,13 @@
 
 from pathlib import Path
 
+import datarobot_genai.nat.datarobot_auth_provider  # noqa: F401 — registers datarobot_api_key
 import httpx
 import pytest
 import respx
 from a2a.types import AgentCapabilities
 from a2a.types import AgentCard
+from datarobot_genai.nat.datarobot_auth_provider import DataRobotAPIKeyAuthProviderConfig
 from nat.builder.context import ContextState
 from nat.builder.workflow_builder import WorkflowBuilder
 from nat.plugins.a2a.client.client_impl import A2AClientFunctionGroup
@@ -29,14 +31,12 @@ from nat.runtime.loader import load_config
 
 import datarobot_genai.dragent.plugins.auth_a2a_client  # noqa: F401 — registers authenticated_a2a_client
 import datarobot_genai.dragent.plugins.okta_a2a_auth  # noqa: F401 — registers okta_cross_app_access
-import datarobot_genai.nat.datarobot_auth_provider  # noqa: F401 — registers datarobot_api_key
 from datarobot_genai.dragent.plugins.auth_a2a_client import AgentCardRegistryLookup
 from datarobot_genai.dragent.plugins.auth_a2a_client import AuthenticatedA2AClientConfig
 from datarobot_genai.dragent.plugins.auth_a2a_client import AuthenticatedA2AClientFunctionGroup
 from datarobot_genai.dragent.plugins.okta_a2a_auth import (
     OAuth2CrossApplicationAccessAuthProviderConfig,
 )
-from datarobot_genai.nat.datarobot_auth_provider import DataRobotAPIKeyAuthProviderConfig
 
 # ---------------------------------------------------------------------------
 # Mock payloads
