+++ tests/dragent/plugins/test_auth_provider.py
@@ -13,13 +13,12 @@
 import os
 from unittest.mock import patch
 
-from nat.authentication.api_key.api_key_auth_provider import APIKeyAuthProvider
-from nat.builder.workflow_builder import WorkflowBuilder
-from pydantic import SecretStr
-
 from datarobot_genai.nat.datarobot_auth_provider import DataRobotAPIKeyAuthProviderConfig
 from datarobot_genai.nat.datarobot_auth_provider import DataRobotMCPAuthProvider
 from datarobot_genai.nat.datarobot_auth_provider import DataRobotMCPAuthProviderConfig
+from nat.authentication.api_key.api_key_auth_provider import APIKeyAuthProvider
+from nat.builder.workflow_builder import WorkflowBuilder
+from pydantic import SecretStr
 
 
 async def test_datarobot_auth_provider():
