+++ tests/dragent/test_context.py
@@ -12,12 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from typing import Any
+
+import jwt
 import pytest
 from nat.builder.context import ContextState
 from nat.data_models.api_server import Request
 from nat.runtime.user_metadata import RequestAttributes
 from starlette.datastructures import Headers
 
+from datarobot_genai.dragent.context import extract_authorization_from_context
+from datarobot_genai.dragent.context import extract_datarobot_headers_from_context
 from datarobot_genai.dragent.context import extract_headers_from_context
 
 
@@ -40,6 +45,92 @@ def nat_context_set_headers():
 
     yield set_headers
     reset_context()
+
+
+@pytest.mark.parametrize(
+    "headers,expected_headers",
+    [
+        (None, {}),
+        ({}, {}),
+        (
+            {"X-DataRobot-Identity-Token": "identity-123", "X-Other-Header": "other-value"},
+            {"x-datarobot-identity-token": "identity-123"},
+        ),
+        (
+            {"X-Untrusted-Foo": "foo-val", "X-Other-Header": "other-value"},
+            {"x-untrusted-foo": "foo-val"},
+        ),
+        (
+            {
+                "X-DataRobot-Identity-Token": "identity-123",
+                "X-Untrusted-Foo": "foo-val",
+                "X-Other-Header": "other-value",
+            },
+            {"x-datarobot-identity-token": "identity-123", "x-untrusted-foo": "foo-val"},
+        ),
+    ],
+)
+def test_extract_datarobot_headers_from_context(headers, expected_headers, nat_context_set_headers):
+    nat_context_set_headers(headers)
+    result = extract_datarobot_headers_from_context()
+    assert result == expected_headers
+
+
+@pytest.mark.parametrize(
+    "headers,expected_authorization_context",
+    [
+        (None, None),
+        ({}, None),
+    ],
+)
+def test_extract_authorization_from_context_empty(
+    headers, expected_authorization_context, nat_context_set_headers
+):
+    nat_context_set_headers(headers)
+    result = extract_authorization_from_context()
+    assert result == expected_authorization_context
+
+
+@pytest.fixture
+def secret_key() -> str:
+    """Return a test secret key for JWT signing."""
+    return "test-secret-key"
+
+
+@pytest.fixture
+def auth_context_data() -> dict[str, Any]:
+    """Return sample authorization context data."""
+    return {
+        "user": {"id": "user123", "name": "Test User", "email": "test@example.com"},
+        "identities": [
+            {
+                "id": "identity123",
+                "type": "user",
+                "provider_type": "datarobot",
+                "provider_user_id": "user123",
+            }
+        ],
+        "metadata": {
+            "endpoint": "https://app.datarobot.com",
+            "account_id": "account456",
+        },
+    }
+
+
+@pytest.fixture
+def auth_token(auth_context_data: dict[str, Any], secret_key: str) -> str:
+    """Generate a valid JWT token from auth context data."""
+    return jwt.encode(auth_context_data, secret_key, algorithm="HS256")
+
+
+def test_extract_authorization_from_context_with_auth_context_data(
+    auth_context_data, secret_key, nat_context_set_headers, auth_token
+):
+    nat_context_set_headers({"X-DataRobot-Authorization-Context": auth_token})
+    result = extract_authorization_from_context(secret_key=secret_key)
+    assert result["user"]["id"] == auth_context_data["user"]["id"]
+    assert result["identities"][0]["id"] == auth_context_data["identities"][0]["id"]
+    assert result["metadata"]["endpoint"] == auth_context_data["metadata"]["endpoint"]
 
 
 @pytest.mark.parametrize(
