--- tests/nat/integration/.env.sample
@@ -1,8 +1,0 @@
-LLM_DEFAULT_MODEL="datarobot/bedrock/anthropic.claude-sonnet-4-5-20250929-v1:0"
-DATAROBOT_ENDPOINT="https://app.datarobot.com/api/v2"
-# Values for buzok-ci-agents@datarobot.com on US prod
-# https://start.1password.com/open/i?a=SNW6QKQ365EQTAV726YE6K4S24&v=jreitiqnr2da7l2itgt5lflm2e&i=qxlkoduit6bi7u3s23n7jik2s4&h=datarobot.1password.com
-DATAROBOT_API_TOKEN=
-SESSION_SECRET_KEY=
-MCP_DEPLOYMENT_ID=
-MEM0_API_KEY=
