--- tests/nat/integration/README.md
@@ -1,8 +1,0 @@
-# NAT integration tests (legacy DRUM path)
-
-These tests exercise NAT end-to-end using the legacy integration path (the way NAT works with DRUM).
-
-## Run
-
-1. Copy values from `.env.sample` into your root `.env` file.
-2. Run `task test-nat-integration`.
