--- tests/nat/integration/fixtures/moderation_real_credentials/moderation_config.yaml
@@ -1,83 +1,0 @@
----
-# Combined guard pack for tests/nat/integration/test_datarobot_moderation_middleware_real_credentials.py
-# (real DATAROBOT_* env). Token / ROUGE / cost OOTB guards plus LLM-gateway-backed OOTB guards.
-#
-# Guardrail catalog: https://github.com/datarobot/moderations/blob/main/docs/GUARDRAILS.md
-#
-# Faithfulness needs CITATION_CONTENT_* columns in the moderation dataframe; without RAG
-# citations the executor logs a config error and does not emit Responses_faithfulness.
-#
-# agent_goal_accuracy is omitted: with llmGateway, dome wraps the client as LlamaIndexLLMWrapper
-# for Ragas, but ragas>=0.3 collections metrics require InstructorLLM (see AgentGoalAccuracyWithoutReference).
-#
-# task_adherence is omitted: DeepEval TaskCompletionMetric may omit task_adherence_score when
-# embeddings are disabled (e.g. "Using MockEmbedding"), so the column is absent from metrics.
-timeout_sec: 180
-guards:
-  - name: Prompt Tokens
-    description: Track the number of tokens associated with the input to the LLM, and/or
-      retrieved text from the vector database.
-    ootb_type: token_count
-    stage: prompt
-    type: ootb
-  - name: Response Tokens
-    description: Track the number of tokens associated with the output from the LLM,
-      and/or retrieved text from the vector database.
-    ootb_type: token_count
-    stage: response
-    type: ootb
-  - name: Response ROUGE-1
-    description: Compare response text overlap with retrieved context using ROUGE-1.
-    ootb_type: rouge_1
-    stage: response
-    type: ootb
-  - name: Response Cost
-    description: Compute response-stage token cost from usage metrics.
-    ootb_type: cost
-    stage: response
-    type: ootb
-    additional_guard_config:
-      cost:
-        currency: USD
-        input_price: 1.0
-        input_unit: 1000
-        output_price: 2.0
-        output_unit: 1000
-  - name: Guideline adherence (gateway)
-    description: Judge whether the response follows the guideline (LLM gateway).
-    ootb_type: agent_guideline_adherence
-    stage: response
-    type: ootb
-    llm_type: llmGateway
-    llm_gateway_model_id: azure/gpt-4o-2024-11-20
-    additional_guard_config:
-      agent_guideline: >-
-        The assistant must answer in plain language and must not refuse to discuss moderation
-        token counting when the user asks about it.
-  - name: Faithfulness (gateway)
-    description: Faithfulness to retrieved citations (LLM gateway; needs RAG citations).
-    ootb_type: faithfulness
-    stage: response
-    type: ootb
-    llm_type: llmGateway
-    llm_gateway_model_id: azure/gpt-4o-2024-11-20
-  - name: Task Adherence (gateway)
-    description: Adherence to the requested task.
-    ootb_type: task_adherence
-    stage: response
-    type: ootb
-    llm_type: llmGateway
-    llm_gateway_model_id: azure/gpt-4o-2024-11-20
-    intervention:
-      action: block
-      message: "Agent did not complete the requested task."
-      conditions:
-        - comparator: lessThan
-          comparand: 0.5
-  - name: Goal Accuracy (gateway)
-    description: Goal accuracy
-    ootb_type: agent_goal_accuracy
-    stage: response
-    type: ootb
-    llm_type: llmGateway
-    llm_gateway_model_id: azure/gpt-4o-2024-11-20
