--- tests/nat/integration/test_agent.py
@@ -1,177 +1,0 @@
-# Copyright 2025 DataRobot, Inc. and its affiliates.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import asyncio
-from concurrent.futures import ThreadPoolExecutor
-from pathlib import Path
-from unittest.mock import ANY
-
-import pytest
-from ag_ui.core import BaseEvent
-from datarobot.core.config import DataRobotAppFrameworkBaseSettings
-
-from datarobot_genai.core.chat import agent_chat_completion_wrapper
-from datarobot_genai.core.chat.completions import convert_chat_completion_params_to_run_agent_input
-from datarobot_genai.core.chat.responses import async_gen_to_sync_thread
-from datarobot_genai.core.chat.responses import (
-    streaming_iterator_to_custom_model_streaming_response,
-)
-from datarobot_genai.core.mcp import MCPConfig
-from datarobot_genai.nat.agent import NatAgent
-from tests.core.chat.test_completions import noop_mcp_tools_factory
-
-
-class Config(DataRobotAppFrameworkBaseSettings):
-    """
-    Finds variables in the priority order of: env
-    variables (including Runtime Parameters), .env, file_secrets, then
-    Pulumi output variables.
-    """
-
-    datarobot_endpoint: str = "https://app.datarobot.com/api/v2"
-    datarobot_api_token: str | None = None
-
-
-@pytest.fixture
-def config():
-    return Config()
-
-
-@pytest.fixture
-def workflow_path():
-    return Path(__file__).parent / "workflow.yaml"
-
-
-@pytest.fixture
-def agent(workflow_path, config):
-    return NatAgent(
-        workflow_path=workflow_path,
-        api_key=config.datarobot_api_token,
-        api_base=config.datarobot_endpoint,
-    )
-
-
-@pytest.fixture
-def workflow_with_mcp_path():
-    return Path(__file__).parent / "workflow_with_mcp.yaml"
-
-
-@pytest.fixture
-def agent_with_mcp(workflow_with_mcp_path, config):
-    mcp_config = MCPConfig()
-    server_config = mcp_config.server_config
-    headers = server_config["headers"] if server_config else None
-    return NatAgent(
-        workflow_path=workflow_with_mcp_path,
-        api_key=config.datarobot_api_token,
-        api_base=config.datarobot_endpoint,
-        forwarded_headers=headers,
-    )
-
-
-async def test_run_method_with_mcp(agent_with_mcp):
-    # You probably need to comment out the disable_env_file conftest fixture to pick up
-    # MCP settings from .env.
-    completion_create_params = {
-        "model": "test-model",
-        "messages": [{"role": "user", "content": "List the projects"}],
-        "environment_var": True,
-    }
-    run_agent_input = convert_chat_completion_params_to_run_agent_input(completion_create_params)
-    streaming_response_iterator = agent_with_mcp.invoke(run_agent_input)
-
-    async for (
-        event,
-        pipeline_interactions,
-        usage,
-    ) in streaming_response_iterator:
-        assert isinstance(event, BaseEvent)
-        assert usage == {
-            "completion_tokens": ANY,
-            "prompt_tokens": ANY,
-            "total_tokens": ANY,
-        }
-    # Final chunk has the total usage and pipeline interactions
-    assert usage["completion_tokens"] > 0
-    assert usage["prompt_tokens"] > 0
-    assert usage["total_tokens"] > 0
-    assert pipeline_interactions
-
-
-async def test_run_method_streaming(agent):
-    # Call the run method with test inputs
-    completion_create_params = {
-        "model": "test-model",
-        "messages": [{"role": "user", "content": "AI"}],
-        "environment_var": True,
-        "stream": True,
-    }
-    run_agent_input = convert_chat_completion_params_to_run_agent_input(completion_create_params)
-    streaming_response_iterator = agent.invoke(run_agent_input)
-
-    async for (
-        event,
-        pipeline_interactions,
-        usage,
-    ) in streaming_response_iterator:
-        assert isinstance(event, BaseEvent)
-        assert usage == {
-            "completion_tokens": ANY,
-            "prompt_tokens": ANY,
-            "total_tokens": ANY,
-        }
-    # Final chunk has the total usage and pipeline interactions
-    assert usage["completion_tokens"] > 0
-    assert usage["prompt_tokens"] > 0
-    assert usage["total_tokens"] > 0
-    assert pipeline_interactions
-
-
-def test_custom_model_streaming_response(agent):
-    # Call the run method with test inputs
-    completion_create_params = {
-        "model": "test-model",
-        "messages": [{"role": "user", "content": "AI"}],
-        "environment_var": True,
-        "stream": True,
-    }
-
-    thread_pool_executor = ThreadPoolExecutor(1)
-    event_loop = asyncio.new_event_loop()
-    thread_pool_executor.submit(asyncio.set_event_loop, event_loop).result()
-
-    result = thread_pool_executor.submit(
-        event_loop.run_until_complete,
-        agent_chat_completion_wrapper(agent, completion_create_params, noop_mcp_tools_factory),
-    ).result()
-
-    streaming_response_iterator = async_gen_to_sync_thread(result, thread_pool_executor, event_loop)
-
-    for response in streaming_iterator_to_custom_model_streaming_response(
-        streaming_response_iterator, model=completion_create_params.get("model")
-    ):
-        result = response.choices[0].delta.content
-        usage = response.usage
-        pipeline_interactions = response.pipeline_interactions
-        if result:
-            assert isinstance(result, str)
-        if usage:
-            assert isinstance(usage.completion_tokens, int)
-            assert isinstance(usage.prompt_tokens, int)
-            assert isinstance(usage.total_tokens, int)
-    # Final chunk has the total usage and pipeline interactions
-    assert usage.completion_tokens > 0
-    assert usage.prompt_tokens > 0
-    assert usage.total_tokens > 0
-    assert pipeline_interactions
