--- tests/nat/integration/workflow_with_mcp.yaml
@@ -1,40 +1,0 @@
-# Copyright 2025 DataRobot, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-functions:
-  completion_tool:
-    _type: chat_completion
-    llm_name: datarobot_llm
-    system_prompt: |
-      You are a helpful assistant.
-
-function_groups:
-  mcp_tools:
-    _type: datarobot_mcp_client
-
-authentication:
-  datarobot_mcp_auth:
-    _type: datarobot_mcp_auth
-
-llms:
-  datarobot_llm:
-    _type: datarobot-llm-gateway
-    model_name: azure/gpt-4o-2024-11-20
-    temperature: 0.0
-
-workflow:
-  _type: per_user_react_agent
-  llm_name: datarobot_llm
-  tool_names:
-    - mcp_tools
-    - completion_tool
