--- tests/nat/integration/workflow_with_streaming_memory.yaml
@@ -1,32 +1,0 @@
-# Copyright 2026 DataRobot, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-memory:
-  mem0_memory:
-    _type: dr_mem0_memory
-
-functions:
-  probe_agent:
-    _type: streaming_memory_probe_agent
-
-workflow:
-  _type: streaming_memory_agent
-  llm_name: unused_for_probe_agent
-  description: "Streaming-memory wrapper integration probe."
-  inner_agent_name: probe_agent
-  memory_name: mem0_memory
-  save_ai_messages_to_memory: false
-  search_params:
-    top_k: 5
-  add_params:
-    async_mode: false
